+++ src/components/loading-spinner/index.ts
@@ -1,0 +1,3 @@
+import LoadingSpinner from './loading-spinner';
+
+export default LoadingSpinner;
