+++ src/components/loading-spinner/loading-spinner.tsx
@@ -1,4 +1,4 @@
-import Header from './header';
+import Header from '../header';
 
 type LoadingSpinnerProps = {
   message?: string;
